--- src/config.js
@@ -1,1 +1,0 @@
-export const API_URL = "https://dhrutham-connect-backend.herokuapp.com";
