+++ src/features/users/userSlice.js
@@ -1,7 +1,7 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import axios from "axios";
 import { logoutButtonClicked } from "../auth/authSlice";
-import { API_URL } from "../../config";
+import { API_URL } from "../../constants";
 
 export const fetchUserProfile = createAsyncThunk(
   "users/userProfile",
